# Conflicts:
#   lib/gitlab/ci/templates/Security/Secret-Detection.gitlab-ci.yml
#   spec/services/ci/create_pipeline_service_spec.rb

+++ ee/app/services/licenses/destroy_service.rb
@@ -4,10 +4,13 @@ module Licenses
   class DestroyService < ::Licenses::BaseService
     extend ::Gitlab::Utils::Override
 
+    DestroyCloudLicenseError = Class.new(StandardError)
+
     override :execute
     def execute
       raise ActiveRecord::RecordNotFound unless license
       raise Gitlab::Access::AccessDeniedError unless can?(user, :destroy_licenses)
+      raise DestroyCloudLicenseError, _('Cloud licenses can not be removed.') if license.cloud_license?
 
       license.destroy
     end
